Rewrite the role-related code
The code for the roles was incorrect according to the official Microsoft documentation.
While the code was functionning, it was unneccessarily complicated, increased the code
to maintain and lead sometimes to errors related to the sharing of database context.
The permissions were handled inefficiently and did not integrate very well with the
.NET framework (including authorization policies)

This commit rewrite and refactors the necessary code for using the roles as they should
according to the documentation.